Grid population functionality added
Wrote a function (createGrid) that uses DOM manipulation to add divs to a container
and change the style of individual grid squares and arrange those divs
equally, using CSS Grid. Another function (getGridRangeValue) gets the
value of the range input element and assigns that to a variable that is
then called as the argument for how many square divs to create.